[app:App] Move `app/App.tsx` to `app/App/index.tsx`
[app:Home] Move `app/Home.tsx` to `app/Home/index.tsx`
[app:Routes] Move `app/AppRoutes.tsx` to `app/Routes/index.tsx`
